fix: Session#getDouble should properly parse valid String, close #4256
Motivation:

Session#getDouble throws a NumberFormatException when the actual value is a String.

Modification:

Fix copy-paste error that was using `Long.parseLong`.
Add more tests covering all the String conversions.

Result:

Double parsing fixed.

+++ gatling-core-java/src/main/java/io/gatling/javaapi/core/Session.java
@@ -189,7 +189,7 @@ public final class Session {
     } else if (value instanceof Double) {
       return (Double) value;
     } else if (value instanceof String) {
-      return Long.parseLong((String) value);
+      return Double.parseDouble((String) value);
     } else {
       throw new ClassCastException(value + " is not an Double: " + value.getClass());
     }
